+++ lib/lightstreamer/stream_connection.rb
@@ -11,6 +11,10 @@ module Lightstreamer
     # @return [String] The control address returned from the server when this stream connection was initiated.
     attr_reader :control_address
 
+    # @return [ProtcolError, RequestError, String] If an error occurs on the stream thread that causes this stream
+    #         to disconnect then the exception or error details will be stored in this attribute.
+    attr_reader :error
+
     # Establishes a new stream connection using the authentication details from the passed session.
     #
     # @param [Session] session The session to create a stream connection for.
@@ -18,37 +22,50 @@ module Lightstreamer
       @session = session
       @queue = Queue.new
 
-      @stream_create_uri = URI.join session.server_url, '/lightstreamer/create_session.txt'
-      @stream_bind_uri = URI.join session.server_url, '/lightstreamer/bind_session.txt'
+      @stream_create_url = URI.join(session.server_url, '/lightstreamer/create_session.txt').to_s
+      @stream_bind_url = URI.join(session.server_url, '/lightstreamer/bind_session.txt').to_s
     end
 
     # Establishes a new stream connection using the authentication details from the session that was passed to
     # {#initialize}. Raises {ProtocolError} or {RequestError} on failure.
     def connect
       return if @thread
-      @session_id = @connection_error = nil
+      @session_id = @error = nil
 
       create_stream_thread
 
       # Wait until the connection result is known
-      until @session_id || @connection_error
+      until @session_id || @error
       end
 
-      raise @connection_error if @connection_error
+      raise @error if @error
+    end
+
+    # Returns whether or not this stream connection is connected.
+    #
+    # @return [Boolean]
+    def connected?
+      !@thread.nil?
     end
 
     # Disconnects this stream connection by shutting down the streaming thread.
     def disconnect
-      return unless @thread
+      if @thread
+        Thread.kill @thread
+        @thread.join
+      end
 
-      Thread.kill @thread
-      @thread.join
-
-      @session_id = @control_address = @connection_error = @thread = nil
+      @session_id = @control_address = @error = @thread = nil
     end
 
-    # Reads the next line of streaming data. This method blocks the calling thread until a line of data is available.
+    # Reads the next line of streaming data. If the streaming thread is alive then this method blocks the calling thread
+    # until a line of data is available. If the streaming thread is not active then any unconsumed lines will be
+    # returned and after that the return value will be `nil`.
+    #
+    # @return [String, nil]
     def read_line
+      return nil if @queue.empty? && @thread.nil?
+
       @queue.pop
     end
 
@@ -59,50 +76,48 @@ module Lightstreamer
         begin
           stream_thread_main
         rescue StandardError => error
-          warn "Lightstreamer: exception in stream thread: #{error}"
-          exit 1
+          @error = error
         end
+
+        @thread = nil
       end
     end
 
     def stream_thread_main
-      http = Net::HTTP.new @stream_create_uri.host, @stream_create_uri.port
-      http.use_ssl = true if @stream_create_uri.port == 443
-
-      connect_stream_and_process_data http, stream_create_post_request
+      connect_stream_and_process_data stream_create_post_request
 
       while @loop
         @loop = false
-        connect_stream_and_process_data http, stream_bind_post_request
+        connect_stream_and_process_data stream_bind_post_request
       end
     end
 
     def stream_create_post_request
-      Net::HTTP::Post.new(@stream_create_uri.path).tap do |request|
-        params = { LS_op2: 'create', LS_cid: 'mgQkwtwdysogQz2BJ4Ji kOj2Bg', LS_user: @session.username,
-                   LS_password: @session.password }
+      params = { LS_op2: 'create', LS_cid: 'mgQkwtwdysogQz2BJ4Ji kOj2Bg', LS_user: @session.username,
+                 LS_password: @session.password }
 
-        params[:LS_adapter_set] = @session.adapter_set if @session.adapter_set
+      params[:LS_adapter_set] = @session.adapter_set if @session.adapter_set
 
-        request.body = URI.encode_www_form(params)
-      end
+      Typhoeus::Request.new @stream_create_url, method: :post, params: params
     end
 
     def stream_bind_post_request
-      Net::HTTP::Post.new(@stream_bind_uri.path).tap do |request|
-        request.body = URI.encode_www_form LS_session: @session_id
-      end
+      Typhoeus::Request.new @stream_bind_url, method: :post, params: { LS_session: @session_id }
     end
 
-    def connect_stream_and_process_data(http, request)
-      buffer = LineBuffer.new
+    def connect_stream_and_process_data(request)
+      @header = StreamConnectionHeader.new
 
-      http.request request do |response|
-        @header = StreamConnectionHeader.new
-        response.read_body do |data|
-          buffer.process data, &method(:process_stream_line)
-        end
+      buffer = LineBuffer.new
+      request.on_body do |data|
+        buffer.process data, &method(:process_stream_line)
       end
+
+      request.on_complete do |response|
+        @error = RequestError.new(response.return_message, response.response_code) unless response.success?
+      end
+
+      request.run
     end
 
     def process_stream_line(line)
@@ -118,7 +133,7 @@ module Lightstreamer
 
       @control_address = @header['ControlAddress']
       @session_id = @header['SessionId']
-      @connection_error = @header.error
+      @error = @header.error
 
       @header = nil
     end